# docs/swagger.yaml
basePath: /
definitions:
...
  models.AuthenticateRequest:
    properties:
      device:
        description: |-
          Device is a unique device identifier
          example: device_123
        type: string
      id:
        description: |-
          ID is either a verification ID (for OTP) or API key ID
          example: 01HNAJ6640M9JRRJFQSZZVE3HH
        type: string
      method:
        description: |-
          Method must be either "otp" or "secret"
          example: otp
        enum:
        - otp
        - secret
        type: string
      signature:
        description: |-
          Signature is a unique device identifier
          example: device_signature_123
        type: string
      token:
        description: |-
          Token is either an OTP code or API key secret
          example: 123456
        type: string
    required:
    - device
    - id
    - method
    - signature
    - token
    type: object
  models.AuthenticateResponse:
    properties:
      token:
        $ref: '#/definitions/models.Token'
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  models.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    type: object
  models.RefreshTokenResponse:
    properties:
      creds:
        type: string
      token:
        $ref: '#/definitions/models.Token'
    type: object
  models.RequestOTPRequest:
    properties:
      phoneNumber:
        description: |-
          PhoneNumber must be in E.164 format (e.g. +60123456789)
          example: +60123456789
        type: string
    required:
    - phoneNumber
    type: object
  models.RequestOTPResponse:
    properties:
      id:
        example: phoneNumberVerification:01HNAJ6640M9JRRJFQSZZVE3HH
        type: string
      message:
        example: message
        type: string
    type: object
...
  models.Token:
    properties:
      accessToken:
        type: string
      accessTokenExpiry:
        description: RefreshToken       string    `json:"refreshToken"` (use ID as
          refresh token)
        type: string
      createdAt:
        type: string
      device:
        type: string
      id:
        type: string
      refreshTokenExpiry:
        type: string
      service:
        type: string
      user:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@infinirewards.io
    name: API Support
    url: http://www.infinirewards.io/support
  description: API for InfiniRewards - A Web3 Loyalty and Rewards Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: InfiniRewards API
  version: "1.0"
paths:
  /auth/authenticate:
    post:
      consumes:
      - application/json
      description: Authenticate user using OTP or API key
      parameters:
      - description: Authentication Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AuthenticateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication successful
          schema:
            $ref: '#/definitions/models.AuthenticateResponse'
        "400":
          description: Invalid request format or validation failed
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Too many attempts
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Authenticate user
      tags:
      - auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh an existing authentication token
      parameters:
      - description: Token Refresh Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/models.RefreshTokenResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh token
      tags:
      - auth
  /auth/request-otp:
    post:
      consumes:
      - application/json
      description: Request a one-time password for authentication
      parameters:
      - description: OTP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            $ref: '#/definitions/models.RequestOTPResponse'
        "400":
          description: Invalid request format or validation failed
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Request OTP
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Enter your bearer token in the format **Bearer <token>**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...